aiservice/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.2.0
transformers==4.35.0
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.24.4
safetensors==0.4.0
huggingface-hub==0.17.3
tokenizers==0.15.0
numpy==1.24.3
scipy==1.11.3
requests==2.31.0
llama-cpp-python==0.2.11